+++ occtax/classes/occtaxSearch.class.php
@@ -112,7 +112,7 @@ class occtaxSearch {
         // Get the number of total records
         if (!$this->recordsTotal && $this->token && !$this->demande) {
             $this->setRecordsTotal();
-            if (in_array($this->name, array('observation', 'brute'))) {
+            if (!in_array($this->name, array('jdd', 'stats', 'taxon'))) {
                 $this->setRecordsExtent();
             }
         }
@@ -498,7 +498,7 @@ class occtaxSearch {
      * Get the number of records returned
     */
     public function getRecordsExtent(){
-        // jLog::log($this->recordsExtent);
+        jLog::log($this->recordsExtent);
         return $this->recordsExtent;
     }